reorganize the tests
