+++ tests/system/load/test_llm.py
@@ -16,7 +16,6 @@ import pandas as pd
 import pytest
 
 import bigframes.ml.llm
-from tests.system import utils
 
 
 @pytest.fixture(scope="session")
@@ -113,46 +112,3 @@ def test_llm_palm_score_params(llm_fine_tune_df_default_index):
         "evaluation_status",
     ]
     assert all(col in score_result_col for col in expected_col)
-
-
-def test_llm_gemini_pro_score(llm_fine_tune_df_default_index):
-    model = bigframes.ml.llm.GeminiTextGenerator(model_name="gemini-pro")
-
-    # Check score to ensure the model was fitted
-    score_result = model.score(
-        X=llm_fine_tune_df_default_index[["prompt"]],
-        y=llm_fine_tune_df_default_index[["label"]],
-    ).to_pandas()
-    utils.check_pandas_df_schema_and_index(
-        score_result,
-        columns=[
-            "bleu4_score",
-            "rouge-l_precision",
-            "rouge-l_recall",
-            "rouge-l_f1_score",
-            "evaluation_status",
-        ],
-        index=1,
-    )
-
-
-def test_llm_gemini_pro_score_params(llm_fine_tune_df_default_index):
-    model = bigframes.ml.llm.GeminiTextGenerator(model_name="gemini-pro")
-
-    # Check score to ensure the model was fitted
-    score_result = model.score(
-        X=llm_fine_tune_df_default_index["prompt"],
-        y=llm_fine_tune_df_default_index["label"],
-        task_type="classification",
-    ).to_pandas()
-    utils.check_pandas_df_schema_and_index(
-        score_result,
-        columns=[
-            "precision",
-            "recall",
-            "f1_score",
-            "label",
-            "evaluation_status",
-        ],
-        index=1,
-    )
